HW_018 Sign Up tests refactor
